Refactor : 담은 기자재 수정 모달 렌더링 이슈 해결
